requirements.txt
numpy
aiohttp
hidapi
livekit
livekit-api
sounddevice
PyAudio